// frontend/src/types/generated/HubSpotDealDiscoveryResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { HubSpotDealDatePropertyOption } from "./HubSpotDealDatePropertyOption";
import type { HubSpotDealPipelineOption } from "./HubSpotDealPipelineOption";

export type HubSpotDealDiscoveryResponse = { configured: boolean, message?: string | null, pipelines: Array<HubSpotDealPipelineOption>, date_properties: Array<HubSpotDealDatePropertyOption>, };