// frontend/src/types/generated/HubSpotDealPipelineMapping.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { HubSpotDealStageMapping } from "./HubSpotDealStageMapping";

export type HubSpotDealPipelineMapping = { pipeline_id: string, stage_mappings: Array<HubSpotDealStageMapping>, started_date_property: string, closed_date_property: string, };