// frontend/src/types/generated/HubSpotDealPipelineMappingResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { HubSpotDealPipelineMapping } from "./HubSpotDealPipelineMapping";

export type HubSpotDealPipelineMappingResponse = { mapping?: HubSpotDealPipelineMapping | null, revision: number | null, };