// frontend/src/types/generated/HubSpotDealPipelineOption.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { HubSpotDealStageOption } from "./HubSpotDealStageOption";

export type HubSpotDealPipelineOption = { pipeline_id: string, label: string, display_order: number, archived: boolean, url?: string | null, stages: Array<HubSpotDealStageOption>, };